Move Convert Into A Class (fixes #12)
Previously, the convert module stored registered conversion functions
inside global state. This made for akward code and also made it
difficult to register multiple conversion functions for the same symbol.

This commit fixes the issue by introducing a new Converter class which
stores all previously global state. For convenience, an instance of this
class `DefaultConverter` is provided as well.  Furthermore, this commit
updates the tests and documentation accordingly.